server/requirements.txt
fastapi>=0.116.1
uvicorn>=0.30.0
python-dotenv>=1.0.1
python-multipart>=0.0.22
starlette>=0.47.2
httpx>=0.28.1
resend>=2.2.0
# Async PG driver (psycopg3) + pool. Replaces psycopg2 — see docs/adr/0002.
psycopg[binary]>=3.2.0
psycopg-pool>=3.2.0
# Schema migrations — see docs/adr/0003.
alembic>=1.13.0
# Optional: error aggregation. Wired only when SENTRY_DSN is set.
sentry-sdk[fastapi]>=2.0.0
# Optional: Prometheus metrics. Wired only when ENABLE_METRICS=true.
prometheus-client>=0.20.0
# Tests
pytest>=8.3.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
requests>=2.32.0
# Dev / lint
ruff>=0.6.0
mypy>=1.10.0
bandit>=1.7.9
pip-audit>=2.7.0